update scripts for batch api

+++ batches/0.create_batch_jsonl.py
@@ -1,7 +1,16 @@
-from pathlib import Path
+import sys
 import os
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+from pathlib import Path
 import json
 import argparse
+import config
+import logging
+import logging_config
+import tiktoken
+logging.disable(level=logging.DEBUG)
+
+SAFETY_CHECK = True
 
 def create_qa_jobs(
     dataset,
@@ -15,12 +24,13 @@ def create_qa_jobs(
     with open(queries_path, 'r') as file:
         queries_info = json.load(file)
     dst = context_path.replace("context.jsonl", "qa_job.jsonl")
-    if os.path.exists(dst):
-        # print(f"{dst} already existed")
-        # return
-        assert False, f"{dst} already existed"
+    assert not os.path.exists(dst), dst
+    assert os.path.exists(os.path.dirname(dst)), dst
+
+    tokenizer = tiktoken.get_encoding("cl100k_base")
     
     job_num = 0
+    input_token_num = 0
     with open(context_path, 'r') as file:
         for l in file:
             context_info = json.loads(l)
@@ -42,29 +52,34 @@ def create_qa_jobs(
                     "max_tokens": max_answer_tokens,
                 }
             }
-            with open(dst, 'a') as file:
-                contents = json.dumps(job) + "\n"
-                file.write(contents)
+            if SAFETY_CHECK == False:
+                with open(dst, 'a') as file:
+                    contents = json.dumps(job) + "\n"
+                    file.write(contents)
             job_num += 1
+            input_token_num += len(tokenizer.encode(prompt))
     
     flag = True
 
-    size_bytes = Path(dst).stat().st_size
-    size_mb = size_bytes / (1024 * 1024)
-    if size_mb > 200:
-        print(f"⚠️ {dst} has {round(size_mb, 2)} MB")
-        flag = False
-    else:
-        print(f"✅ {dst} has {round(size_mb, 2)} MB")
+    if SAFETY_CHECK == False:
+        size_bytes = Path(dst).stat().st_size
+        size_mb = size_bytes / (1024 * 1024)
+        if size_mb > 200:
+            print(f"⚠️ {dst} has {round(size_mb, 2)} MB")
+            flag = False
+        else:
+            print(f"✅ {dst} has {round(size_mb, 2)} MB")
 
-    if job_num > 100000:
-        print(f"⚠️ {dst} has {job_num} requests")
-        flag = False
-    else:
-        print(f"✅ {dst} has {job_num} requests")
+        if job_num > 100000:
+            print(f"⚠️ {dst} has {job_num} requests")
+            flag = False
+        else:
+            print(f"✅ {dst} has {job_num} requests")
 
-    if flag == False:
-        os.remove(dst)
+        if flag == False:
+            os.remove(dst)
+
+    print(f"✅ ${round((input_token_num * 0.075 / 1000000) + (job_num * 100 * 0.3 / 100000), 6)} costs for {dst}")
 
     return flag    
 
@@ -83,9 +98,8 @@ def create_llmjudge_jobs(
     with open(queries_path, 'r') as file:
         queries = json.load(file)
     dst = answer_path.replace("answer.jsonl", "rating_job.jsonl")
-    if os.path.exists(dst):
-        assert False, f"{dst} already existed"
-        # return
+    assert not os.path.exists(dst), dst
+    assert os.path.exists(os.path.dirname(dst)), dst
 
     with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), "llm_judge_prompt.txt"), 'r') as file:
         prompt_template = file.read().strip()
@@ -100,11 +114,16 @@ def create_llmjudge_jobs(
             answers.append(json.loads(l))
 
     assert len(answers) == len(queries)
-    
+
+    tokenizer = tiktoken.get_encoding("cl100k_base")
+
     job_num = 0
+    input_token_num = 0
     for answer_info, query_info in zip(answers, queries):
         assert answer_info["id"] == query_info["id"]
         my_answer = answer_info["answer"]
+        if my_answer == None:
+            continue
         gold_answers = query_info["answer"]
         for ga_id, gold_answer in enumerate(gold_answers):
             prompt = prompt_template.replace("{query}", query_info['query']).replace("{reference}", gold_answer).replace("{answer}", my_answer)
@@ -122,86 +141,79 @@ def create_llmjudge_jobs(
                     "max_tokens": max_answer_tokens,
                 }
             }
-            with open(dst, 'a') as file:
-                contents = json.dumps(job) + "\n"
-                file.write(contents)
+            if SAFETY_CHECK == False:
+                with open(dst, 'a') as file:
+                    contents = json.dumps(job) + "\n"
+                    file.write(contents)
             job_num += 1
+            input_token_num += len(tokenizer.encode(prompt))
 
     flag = True
 
-    size_bytes = Path(dst).stat().st_size
-    size_mb = size_bytes / (1024 * 1024)
-    if size_mb > 200:
-        print(f"⚠️ {dst} has {round(size_mb, 2)} MB")
-        flag = False
-    else:
-        print(f"✅ {dst} has {round(size_mb, 2)} MB")
+    if SAFETY_CHECK == False:
+        size_bytes = Path(dst).stat().st_size
+        size_mb = size_bytes / (1024 * 1024)
+        if size_mb > 200:
+            print(f"⚠️ {dst} has {round(size_mb, 2)} MB")
+            flag = False
+        else:
+            print(f"✅ {dst} has {round(size_mb, 2)} MB")
 
-    if job_num > 100000:
-        print(f"⚠️ {dst} has {job_num} requests")
-        flag = False
-    else:
-        print(f"✅ {dst} has {job_num} requests")
+        if job_num > 100000:
+            print(f"⚠️ {dst} has {job_num} requests")
+            flag = False
+        else:
+            print(f"✅ {dst} has {job_num} requests")
 
-    if flag == False:
-        os.remove(dst)
+        if flag == False:
+            os.remove(dst)
+
+    print(f"✅ ${round((input_token_num * 0.075 / 1000000) + (job_num * 100 * 0.3 / 100000), 6)} costs for {dst}")
 
     return flag    
 
 
 if __name__ == "__main__":
-    NODE_EMBEDDING_MODEL_LIST = ["sbert", "dpr", "te3small", "bm25"]
-    CONTEXT_LEN_RATIO_LIST = [0.05, 0.1, 0.15, 0.2]
-    CONTEXT_CONFIG_LIST = (
-        # end-to-end
-        [("vanilla", None, nem, None, None, None, r) for nem in NODE_EMBEDDING_MODEL_LIST for r in CONTEXT_LEN_RATIO_LIST] + 
-        [("raptor", None, nem, None, None, None, r) for nem in NODE_EMBEDDING_MODEL_LIST for r in CONTEXT_LEN_RATIO_LIST] + 
-        [("sht", None, nem, True, True, True, r) for nem in NODE_EMBEDDING_MODEL_LIST for r in CONTEXT_LEN_RATIO_LIST]
-        # # ablation on SHT
-        # [("sht", sht_type, "sbert", True, True, True, 0.15) for sht_type in SHT_TYPE_LIST] + 
-        # # ablation on HI
-        # [("sht", None, "sbert", False, True, True, 0.15), ("sht", None, "sbert", True, False, True, 0.15), ("sht", None, "sbert", False, False, True, 0.15)] +
-        # # ablation on CI
-        # [("sht", None, "sbert", True, True, False, 0.15)]
-    )
+    parser = argparse.ArgumentParser()
 
+    parser.add_argument("--dataset", type=str, required=True, choices=["qasper", "civic", "contract", "finance"])
+    parser.add_argument("--job", type=str, required=True, choices=["qa", "llmjudge"])
+    parser.add_argument("--safety-check", action='store_true', help="Whether to check the context configures in config.py before continue")
+    args = parser.parse_args()
 
-    batch_num = 0
-    for dataset in ["qasper"]:
-        for method, sht_type, node_embedding_model, embed_hierarchy, context_hierarchy, use_raw_chunks, context_len_ratio in CONTEXT_CONFIG_LIST:
-            context_jsonl_path = f"/home/ruiying/SHTRAG/data/{dataset}"
-            if method != "sht":
-                context_jsonl_path = os.path.join(context_jsonl_path, "baselines")
-                context_folder_suffix = f"raptor{int(method=='raptor')}"
-            else:
-                assert embed_hierarchy != None
-                assert isinstance(embed_hierarchy, bool)
-                context_folder_suffix = f"h{int(embed_hierarchy == True)}"
-            if sht_type != None:
-                context_jsonl_path = os.path.join(context_jsonl_path, sht_type)
-            query_embedding_model = node_embedding_model
-            context_jsonl_path = os.path.join(
-                context_jsonl_path,
-                f"{node_embedding_model}.gpt-4o-mini.c100.s100",
-                f"{query_embedding_model}.cosine.{context_folder_suffix}",
-                "o0" if method != "sht" else "l1.h1",
-                f"context{context_len_ratio}",
-                "context.jsonl"
-            )
+    SAFETY_CHECK = bool(args.safety_check)
 
-            assert os.path.exists(context_jsonl_path), context_jsonl_path
-            print(context_jsonl_path)
+    if args.job == "llmjudge":
+        assert args.dataset == "qasper", "Only Qasper need LLM judge."
 
-            if context_jsonl_path == "/home/ruiying/SHTRAG/data/civic/baselines/bm25.gpt-4o-mini.c100.s100/bm25.cosine.raptor0/o0/context0.05/context.jsonl":
-                continue
-
-            batch_num += 1
-            
+    job_num = 0
+    check_context_list = input(f"Check the context configures in config.py before continue (dataset={args.dataset}, job={args.job}, safety_check={SAFETY_CHECK})... [y/n]")
+    if check_context_list.lower() != 'y':
+        print("Exit")
+        exit(0)
+    else:
+        print("Continue...")
+    
+    for context_config in config.CONTEXT_CONFIG_LIST:
+        context_jsonl_path = config.get_config_jsonl_path(args.dataset, context_config)
+        assert os.path.exists(context_jsonl_path)
+        logging.info(f"Creating {args.job} jobs for {context_jsonl_path}...")
+        job_num += 1
+        if args.job == "qa":
             create_qa_jobs(
-                dataset=dataset,
+                dataset=args.dataset,
                 context_path=context_jsonl_path,
                 max_answer_tokens=100,
                 model="gpt-4o-mini"
             )
-
-    print(batch_num)
+        else:
+            assert args.job == "llmjudge"
+            answer_path = context_jsonl_path.replace("context.jsonl", "answer.jsonl")
+            assert os.path.exists(answer_path), answer_path
+            create_llmjudge_jobs(
+                answer_path=answer_path,
+                max_answer_tokens=100,
+                model="gpt-4o-mini"
+            )
+    
+    print(f"Created {job_num} jobs on local machine!")
